Move rate limit module to utils and generate deprecation warnings
